requirements.txt
# requirements.txt
requests==2.28.2
beautifulsoup4==4.11.2
selenium==4.8.2
webdriver-manager==3.8.5
pandas==1.5.3
nltk==3.8.1
googletrans==4.0.0-rc1
python-dotenv==1.0.0
SQLAlchemy==2.0.4
fastapi==0.95.0  # Se quiser criar uma API
uvicorn==0.21.1  # Para servir a API